pyproject.toml
[tool.poetry]
name = "diplom-todolist"
version = "0.1.0"
description = "Дипломный проект"
authors = ["Dmitriy Lahmenev <lahmenev@gmail.com>"]
readme = "README.md"
packages = [{include = "diplom_todolist"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
social-auth-app-django = "^5.0.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
ansible-vault-win = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tools.isort]
py_version = 310
profile = 'django'
line_length =120 # длина строики
multi_line_output = 2 # отступ между классами
known_first_party = ['todolist_project', 'core'] # наши написанные приложения  для проверки


[tool.flake8]
max-line-length = 120
per-file-ignores = """
    */migrations/*: E501,
"""
